Done with Technical Class Notes by pdf
What is web responsive design?
Responsive web design (RWD) is a web development approach that creates dynamic changes to the appearance of a website, depending on the screen size and orientation of the device being used to view it.is defined simply as building a website that works on all devices and screen sizes. It is and always has been about creating sites that work on any device and any screen size. These designs are targeted to meet the needs of users of all kinds
What are the Three Main Components of a Responsive Website Design?

Responsive web design is divided into three main components: the media query, the web browser, and the responsive web interface itself.

Media Queries
Media queries are supported by the latest web browsers and work by creating media queries. They are called “media queries” and can be used in various ways, such as in connection with the layout of a page or as part of the content of an application. 

They also permit designers to fabricate various formats utilizing similar HTML archives by specifically serving templates dependent on the client specialist’s highlights, like the browser’s window’s size.

Web  Browser
Since websites can contain many images, it is essential to keep these images fluid too. The thought behind liquid pictures is that you convey concepts at the greatest size they will utilize. In responsive web pages, designers don’t announce the stature and width in your code; instead, they let the browsers resize the pictures on a case by case basis while utilizing CSS to manage their relative size. It’s an extraordinary and straightforward procedure to resize images correctly.

Website interface (JavaScript, HTML, and CSS) 
Many mobile devices are touchscreen nowadays, which requires mindfulness of the size of the interactive elements within interfaces.

Aside from the flexibility of images, usage of native controls such as keyboards and drop-out menus should be optimized so it can provide a great experience to its users, whether it’s on mobile or desktop.